# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPosts": {
      "get": {
        "tags": [
          "BlogPosts"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogPostDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogPostDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogPostDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BlogPosts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/BlogPosts/{id}": {
      "get": {
        "tags": [
          "BlogPosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BlogPosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPostDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BlogPosts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlogPostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CreateBlogPostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateBlogPostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}